+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+- 2021-12-01
+  - Changed the type of param of `BezierPath.scale` to XYCoords.
+  - Added function `BezierPath.scaleXY` for non uniform scaling.
+  - Changed the type of param of `Vertex.scale` to XYCoords.
+  - Added function `Vertex.scaleXY` for non uniform scaling.
 - 2021-11-24
   - Speeding up the demo page with async script execution.
   - Finalizing the demo page for the new `PBText` element (note: it has this name because `Text` already exists).